apps/repair-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
supabase==2.0.0
httpx==0.25.2
docker==7.0.0
anthropic==0.25.0
PyGithub==2.1.1
redis==5.0.1
python-multipart==0.0.6
sentry-sdk==1.39.2
structlog==24.1.0
